requirements.txt
branca==0.7.2
colorama==0.4.6
Flask==3.0.3
folium==0.17.0
geopandas==1.0.1
Pillow==10.4.0
PyQt5==5.15.11
PyQt5_sip==12.15.0
python_dateutil==2.9.0.post0
Requests==2.32.3
selenium==4.24.0
Shapely==2.0.6
tabulate==0.9.0
utils==1.0.2
webdriver_manager==4.0.2
